Move patching helpers to helpers.ts
